Backtrace (moderate)
For debugging it is often useful to have a backtrace: a list of the function
calls on the stack above the point at which the error occurred. To help
with backtraces, the compiler generates machine code that maintains a stack
frame on the stack corresponding to each function in the current call chain.
Each stack frame consists of the return address and a "frame pointer" to the
caller's stack frame. Register s0 contains a pointer to the current stack
frame (it actually points to the the address of the saved return address on
the stack plus 8). Your backtrace should use the frame pointers to walk up
the stack and print the saved return address in each stack frame.

Implement a backtrace() function in kernel/printf.c. Insert a call to
this function in sys_sleep, and then run bttest, which calls sys_sleep.
Your output should be a list of return addresses with this form
(but the numbers will likely be different):

    backtrace:
    0x0000000080002cda
    0x0000000080002bb6
    0x0000000080002898

After bttest exit qemu. In a terminal window: run addr2line -e
kernel/kernel (or riscv64-unknown-elf-addr2line -e kernel/kernel)
and cut-and-paste the addresses from your backtrace, like this:
    $ addr2line -e kernel/kernel
    0x0000000080002de2
    0x0000000080002f4a
    0x0000000080002bfc
    Ctrl-D

You should see something like this:
    kernel/sysproc.c:74
    kernel/syscall.c:224
    kernel/trap.c:85

Some hints:

Add the prototype for your backtrace() to kernel/defs.h so that
you can invoke backtrace in sys_sleep.
The GCC compiler stores the frame pointer of the currently executing
function in the register s0. Add the following function to kernel/riscv.h:
static inline uint64
r_fp()
{
  uint64 x;
  asm volatile("mv %0, s0" : "=r" (x) );
  return x;
}
and call this function in backtrace to read the current frame pointer.
r_fp() uses in-line assembly to read s0.
These lecture notes have a picture of the layout of stack frames.
Note that the return address lives at a fixed offset (-8) from the frame
pointer of a stackframe, and that the saved frame pointer lives at
fixed offset (-16) from the frame pointer.
Your backtrace() will need a way to recognize that it has seen the last
stack frame, and should stop. A useful fact is that the memory allocated
for each kernel stack consists of a single page-aligned page, so that
all the stack frames for a given stack are on the same page. You can use
PGROUNDDOWN(fp) (see kernel/riscv.h) to identify the page that a frame
pointer refers to.
Once your backtrace is working, call it from panic in kernel/printf.c
so that you see the kernel's backtrace when it panics.